# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Swagger documentation for the OpenCost costmodel HTTP API.",
    "title": "OpenCost API",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "paths": {
...
    "/kapis/costwise.wiztelemetry.io/v1alpha1/allocation": {
      "get": {
        "description": "返回成本分配明细数据，支持聚合、过滤、idle、共享 idle、比例资产成本和 CSV 导出。",
        "tags": [
          "Allocation"
        ],
        "summary": "查询成本分配明细",
        "parameters": [
          {
            "description": "时间窗口。必填。",
            "name": "window",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "聚合维度，多个值使用英文逗号分隔。",
            "name": "aggregate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "是否包含 idle 成本。",
            "name": "includeIdle",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "是否按节点级别计算 idle。",
            "name": "idleByNode",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "是否共享负载均衡成本。",
            "name": "sharelb",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "是否纳入比例资产资源成本。",
            "name": "includeProportionalAssetResourceCosts",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "是否附带聚合元数据。",
            "name": "includeAggregatedMetadata",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "是否回摊 idle 成本。",
            "name": "shareIdle",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "累计粒度。",
            "name": "accumulate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "显式累计粒度，覆盖 accumulate。",
            "name": "accumulateBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "查询步长。",
            "name": "step",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "分配过滤条件。",
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "返回格式；传 csv 导出 CSV。",
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          }
        }
      }
    },
    "/kapis/costwise.wiztelemetry.io/v1alpha1/allocation/autocomplete": {
      "get": {
        "description": "兼容 Kubecost 的 allocation autocomplete 接口。当前支持 label、label[<key>]、annotation、annotation[<key>]、cluster、node、namespace、pod、container、controller、controllerKind、providerID、service。node 返回 cluster/node，pod 返回 cluster/namespace/pod，container 返回 cluster/namespace/pod/container，service 返回 cluster/namespace/service。",
        "tags": [
          "Allocation"
        ],
        "summary": "查询分配字段自动补全候选项",
        "parameters": [
          {
            "description": "时间窗口。必填。",
            "name": "window",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "字段名。支持 label、label[<key>]、annotation、annotation[<key>]、cluster、node、namespace、pod、container、controller、controllerKind、providerID、service。",
            "name": "field",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "搜索关键字，按包含关系过滤候选项。",
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "分配过滤条件。",
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/costmodel.Response"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "costmodel.Response": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "data": {},
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "warning": {
            "type": "string"
          }
        }
      }
    }
  }
}